ci: follow dotenv artifact rules
The [`dotenv` artifact for `prepare-nightly` is invalid](https://gitlab.com/fintechstudios/davical-docker/-/jobs/3497020806). It doesn't list the exact reason, but according to [the rules for `dotenv` artifacts](https://docs.gitlab.com/ee/ci/yaml/artifacts_reports.html#artifactsreportsdotenv):

- The file cannot have empty lines
- Values cannot have spaces

So, this corrects the .env file to follow these rules. This required moving `DOCKER_BUILD_FLAGS` and `BUILD_PREFIX` out of the `.env`, so they are now created in the `.gitlab-ci.yaml`