ci: remove quotation marks from dotenv file
Despite some reports that GitLab dotenv files can use quotation marks around the values, this appears to _not_ be the case, as the [quotation marks are being included in the variable values](https://gitlab.com/fintechstudios/davical-docker/-/jobs/3497342262). Luckily, they should not be necessary for any of the included variables, so this removes them.