ci: fix release SHA helper
The GitLab /archive/ endpoint will return a different bundle when a commit hash is specified. This means that if "master" is used to download the archive, and then the same commit hash that download is used to download a separate archive, the sha512sum of both archives will NOT match.

As a result, the nightly build fails, since the sha512sum is computed by downloading the "master" archive, but the image is built using the commithash. This resolves the issue by explicitly fetching the latest commit hash for master BEFORE computing the shasum of the archive.