ci: fix release SHA helper (#12)
The GitLab `/archive/` endpoint will return a different bundle when a commit hash is specified. This means that if "master" is used to download the archive, and then the same commit hash from that download is used to download a separate archive, the shasum of both archives will not match.

As a result, the nightly build fails, since the shasum is computed by downloading the "master" archive, but the image is built using the commithash. This resolves the issue by explicitly fetching the latest commit hash for master BEFORE computing the shasum of the archive.