ci: use short commit hash for versioning
The nightly image uses the DAViCAl and AWL commit hashes for versioning, but having the full hash for both (as well as build information) in the image tag causes it to [exceed the maximum tag character length of 128](https://gitlab.com/fintechstudios/davical-docker/-/jobs/3498334697).

This changes to use the short commit hash (the first 8 characters) for versioning, which should prevent this issue from occurring.